+++ apps/web/src/features/menu/ui/SettingModal.tsx
@@ -3,15 +3,13 @@ import React, { useEffect, useRef, useState } from 'react';
 import { cn } from '@soup/utils';
 import { Button } from '@soup/design-system';
 
+import { Modal } from '~/shared/ui';
 import { MODAL, SETTING_ITEM } from '~/shared/constants';
 import { useModal, useModalState } from '~/shared/hooks';
 import type { ModalRef, SettingItem } from '~/shared/types';
-import { Modal } from '~/shared/ui';
-import {
-  ProfileSetting,
-  AlarmSetting,
-  WithdrawSetting,
-} from '~/widgets/modal/ui/setting';
+import ProfileSetting from './ProfileSetting';
+import AlarmSetting from './AlarmSetting';
+import WithdrawSetting from './WithdrawSetting';
 
 export default function SettingModal() {
   const { isOpen: openSetting } = useModalState({ key: MODAL.SETTING });
